Fixed and completed sentiment analysis example
