+++ src/Examples/CrowdsourcingWithWords/BCCWords.cs
@@ -165,17 +165,4 @@ namespace CrowdsourcingWithWords
             return posteriors;
         }
     }
-
-    /// <summary>
-    /// BCCWords posterior object.
-    /// </summary>
-    [Serializable]
-    public class BCCWordsPosteriors : BCCPosteriors
-    {
-        /// <summary>
-        /// The Dirichlet posteriors of the word probabilities for each true label value.
-        /// </summary>
-        public Dirichlet[] ProbWordPosterior;
-
-    }
 }
