+++ src/Examples/CrowdsourcingWithWords/CrowdsourcingWithWords.cs
@@ -12,21 +12,21 @@
 
 namespace CrowdsourcingWithWords
 {
-    using System;
-    using System.Collections.Generic;
-    using System.IO;
+	using System;
+	using System.Collections.Generic;
+	using System.IO;
 
-    class CrowdsourcingWithWords
+	class CrowdsourcingWithWords
     {
         /// <summary>
         /// Main method to run the crowdsourcing experiments presented in Simpson et.al (WWW15).
         /// </summary>
         public static void Main()
         {
-            var data = Datum.LoadData(Path.Combine("Data", "labels.tsv"));
+			var data = Datum.LoadData(Path.Combine("Data", "weatherTweets.tsv"));
 
-            // Run model and get results
-            var VocabularyOnSubData = ResultsWords.BuildVocabularyOnSubdata((List<Datum>)data);
+			// Run model and get results
+			var VocabularyOnSubData = ResultsWords.BuildVocabularyOnSubdata((List<Datum>)data);
 
             BCCWords model = new BCCWords();
             ResultsWords resultsWords = new ResultsWords(data, VocabularyOnSubData);
@@ -45,74 +45,6 @@ namespace CrowdsourcingWithWords
 
             Console.WriteLine("Done.  Press enter to exit.");
             Console.ReadLine();
-        }
-    }
-
-    /// <summary>
-    /// This class represents a single datum, and has methods to read in data.
-    /// </summary>
-    public class Datum
-    {
-        /// <summary>
-        /// The worker id.
-        /// </summary>
-        public string WorkerId;
-
-        /// <summary>
-        /// The task id.
-        /// </summary>
-        public string TaskId;
-
-        /// <summary>
-        /// The worker's label.
-        /// </summary>
-        public int WorkerLabel;
-
-        /// <summary>
-        /// The task's gold label (optional).
-        /// </summary>
-        public int? GoldLabel;
-
-        /// <summary>
-        /// The body text of the document (optional - only for text sentiment labelling tasks).
-        /// </summary>
-        public string BodyText;
-
-        /// <summary>
-        /// Loads the data file in the format (worker id, task id, worker label, ?gold label).
-        /// </summary>
-        /// <param name="filename">The data file.</param>
-        /// <param name="maxLength"></param>
-        /// <returns>The list of parsed data.</returns>
-        public static IList<Datum> LoadData(string filename, int maxLength = short.MaxValue)
-        {
-            var result = new List<Datum>();
-            using (var reader = new StreamReader(filename))
-            {
-                string line;
-                while ((line = reader.ReadLine()) != null && result.Count < maxLength)
-                {
-                    var strarr = line.Split('\t');
-                    int length = strarr.Length;
-
-                    var datum = new Datum
-                    {
-                        WorkerId = strarr[0],
-                        TaskId = strarr[1],
-                        WorkerLabel = int.Parse(strarr[2]),
-                        BodyText = strarr[3]
-                    };
-
-                    if (length >= 5)
-                        datum.GoldLabel = int.Parse(strarr[4]);
-                    else
-                        datum.GoldLabel = null;
-
-                    result.Add(datum);
-                }
-            }
-
-            return result;
         }
     }
 }
