+++ src/Examples/CrowdsourcingWithWords/CrowdsourcingWithWords.csproj
@@ -74,5 +74,8 @@
     <None Update="Data\stopwords.txt">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
+    <None Update="Data\weatherTweets.tsv">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
   </ItemGroup>
 </Project>
