+++ src/Examples/CrowdsourcingWithWords/TFIDFClass.cs
@@ -301,19 +301,4 @@ namespace CrowdsourcingWithWords
 
         #endregion
     }
-
-    /// <summary>
-    /// This is a placeholder class.  Replace with your own stemmer.
-    /// </summary>
-    public class EnglishWord
-    {
-        public readonly string Original;
-        public readonly string Stem;
-
-        public EnglishWord(string word)
-        {
-            Original = word;
-            Stem = word;
-        }
-    }
 }
